Fix 404/500 issues on tools/scummvm for both systems and images
**Issues Fixed:**
- tools and scummvm systems show 404
- scummvm and tools have images in a non-standard location and also gives 404
- Dates on certain tools scripts have dates which issue a 500

**Root Cause and Logic to Address Issues**:
- Tools and scummvm have non-standard locations for their ‘roms_folder’.
	- **Incorrect assumption in code**: The main roms folder (/storage/roms) + system name can be used to find the roms directory for a given system
	- **Fix logic**: look up system folder in es_systems.cfg.  NOTE: a simplification possible by using this lookup - can use 'name' of system to identify it (no need for 'folder')
- Tools and scummvm  have non-standard locations for their ‘images’
	- **Bad assumption in code**: The system roms folder + “images/”+ image name can be used to find an image
	- **Fix logic**: use path from gamedata.xml and don’t derive images path
          - Also fix upload to create 'images' folder if it does not exist
          - Fix image logic to not issue 500 on file size if file doesn't exit (gamedata.xml references missing file)
- Tools provides dates like ‘2020’ which are not parsed
	- **Fix logic**: if a date cannot be parsed for formatting, just display it.

**Fix Overview**
Most of the above fixes are simple to implement.  However, getting the system roms directory from es_systems.cfg is not always obvious as it is often needed on a per game basis in contexts where es_systems.cfg has not been parsed. Parsing it for each game (for example) would be a huge overhead.

The proposed solution is to lazily cache the systems.cfg object in memory and use it across all APIs.  This is the most controversial part of this PR as there is no existing pattern to follow for this in the code.  A few thoughts as to why this is ok:
- the file is read only and the in memory object is not updated so it should be ok to use concurrently.
- In the event the server were to be multithreaded, each thread would just cache es_systems.cfg so it should work without additional logic.

NOTE on memory: Caching the es_systems.cfg object does potentially use a bit more memory, but it should be relatively small.  If this were to be an issue, parsing and keeping a dictionary with only the system paths would be a good option.